+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Method for display the day of the week in charts
+- Method for using the component in TD
 
 ## 9.16.1 - 2021-01-11